requirements.txt
PyMuPDF 
azure-storage-blob 
azure-cosmos 
openai
cryptography==41.0.7